requirements.txt
# python==3.5
django==1.9.6
django-environ==0.4.0
django-crispy-forms==1.6.0
django-modeltranslation==0.11
factory-boy==2.7.0
pillow==3.2.0
